Only place method annotations if the declaringStructure is the relevant class
The declaringClass is not the proper property to check as it will
also point to the class being checked if it is originating from a
trait. If that is the case, the annotation shouldn't be shown in
the file of the relevant class as the override itself isn't present
there either (it is instead present in the file of the trait).